+++ .files/AssemblyInfo.Version.cs
@@ -7,17 +7,8 @@ using Platform.Support;
 using Platform.Support.Attributes;
 #endif
 
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version 
-//      Build Number
-//      Revision
-//
-
-[assembly: AssemblyVersion("3.1.0.*")]
-[assembly: AssemblyFileVersion("3.1.0")]
-//[assembly: AssemblyInformationalVersion("3.1.0-master")]
+[assembly: AssemblyVersion("3.1.11.*")]
+[assembly: AssemblyFileVersion("3.1")]
 
 #if (!DEBUG)
 [assembly: ProductLevel(ProductLevels.RTW)]