+++ Presentation.Forms/Controls/AutoHeightLabel.cs
@@ -1,0 +1,40 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Windows.Forms;
+
+namespace Platform.Presentation.Forms.Controls
+{
+    [ToolboxBitmap(typeof(Label)), DebuggerNonUserCode()]
+    public class AutoHeightLabel : Label
+    {
+        // Methods
+        protected override void OnSizeChanged(EventArgs e)
+        {
+            base.OnSizeChanged(e);
+            this.ResetHeight();
+        }
+
+        private void ResetHeight()
+        {
+            Size preferredSize = this.GetPreferredSize(base.Size);
+            base.Size = new Size(base.Width, preferredSize.Height);
+        }
+
+
+        // Properties
+        public override string Text
+        {
+            get { return base.Text; }
+            set
+            {
+                base.Text = value;
+                this.ResetHeight();
+            }
+        }
+
+    }
+}